- normal death = 0.5 - player kill = 1

+++ src/main/kotlin/net/fabricmc/example/server/commands/LivesCommand.kt
@@ -13,7 +13,7 @@ val livesCommand = command("lives") {
         val uuid = source.player?.uuid ?: return@runs
         source.player?.sendText(literalText("Deine Lebenspunkte sind ${getConfigValue(uuid)}"))
     }
-    argument<Int>("amount") { amount ->
+    argument<Double>("amount") { amount ->
         argument<String>("player") { target ->
             runs {
                 val server = source.player?.server ?: return@runs
